+++ Runtime/Simulation/Subsystems/EnvironmentProbe/SimulationEnvironmentProbeDiscoverer.cs
@@ -38,6 +38,7 @@ namespace UnityEngine.XR.Simulation
         int m_DiscoveredAutomaticCount;
         bool m_IsSceneInitialized;
         bool m_IsStarted;
+        bool m_AutomaticPlacementEnabled;
 
         public IReadOnlyList<XREnvironmentProbe> added => m_Added;
         public IReadOnlyList<XREnvironmentProbe> updated => m_Updated;
@@ -46,6 +47,12 @@ namespace UnityEngine.XR.Simulation
         public bool IsReady => m_IsSceneInitialized && m_IsStarted;
 
         public bool HasAnyChanges => m_Added.Count > 0 || m_Updated.Count > 0 || m_Removed.Count > 0;
+
+        internal bool automaticPlacementEnabled
+        {
+            get { return m_AutomaticPlacementEnabled; }
+            set { m_AutomaticPlacementEnabled = value; }
+        }
 
         public void Dispose()
         {
@@ -273,7 +280,7 @@ namespace UnityEngine.XR.Simulation
         void ScanForUndiscoveredProbes()
         {
             var automaticProbesCount = m_AutomaticProbes.Count;
-            if (m_DiscoveredAutomaticCount == automaticProbesCount)
+            if (m_DiscoveredAutomaticCount == automaticProbesCount || !m_AutomaticPlacementEnabled)
                 return;
 
             GeometryUtility.CalculateFrustumPlanes(m_OriginCameraData.camera, s_CameraPlanes);