+++ Runtime/Simulation/Subsystems/EnvironmentProbe/SimulationEnvironmentProbeSubsystem.cs
@@ -18,9 +18,15 @@ namespace UnityEngine.XR.Simulation
             EnvironmentProbeParams m_EnvironmentProbeParams;
             float m_LastUpdateTime;
 
-            public override bool automaticPlacementRequested { get; set; }
+            public override bool automaticPlacementRequested {
+                get { return m_ProbeDiscoverer?.automaticPlacementEnabled ?? false; }
+                set {
+                    if (m_ProbeDiscoverer != null)
+                        m_ProbeDiscoverer.automaticPlacementEnabled = value;
+                }
+            }
 
-            public override bool automaticPlacementEnabled => true;
+            public override bool automaticPlacementEnabled => automaticPlacementRequested;
 
             public override bool environmentTextureHDREnabled => true;
 
@@ -29,6 +35,7 @@ namespace UnityEngine.XR.Simulation
             protected override bool TryInitialize()
             {
                 m_ProbeDiscoverer = new SimulationEnvironmentProbeDiscoverer();
+                m_ProbeDiscoverer.automaticPlacementEnabled = automaticPlacementRequested;
                 m_EnvironmentProbeParams = XRSimulationRuntimeSettings.Instance.environmentProbeDiscoveryParams;
                 return base.TryInitialize();
             }